test: move test to internal/test
Fixes #521
